Changed to the shared-db hook to allow creation of multiple users/databases in the same hook execution context.
Todo this this hook has been re-implemented in python; existing interface
functionality remains unchanged but the hook will also now accept requests
for multiple username/databases at the same time. This was implemented to
allow quantum and nova services to be hosted in the same charm without the
requirement for additional relations or subordinates.

Removed print of root password in juju log.